+++ hypertunity/__init__.py
@@ -1,3 +1,4 @@
 from .optimisation import *
 from .reports import *
 from .scheduling import *
+from .trial import *