+++ hypertunity/scheduling/tests/test_scheduler.py
@@ -13,7 +13,7 @@ def square(sample: ht.Sample) -> ht.EvaluationScore:
     return ht.EvaluationScore(sample["x"] ** 2)
 
 
-def _run_jobs(jobs):
+def run_jobs(jobs):
     with Scheduler(n_parallel=2) as scheduler:
         scheduler.dispatch(jobs)
         results = scheduler.collect(n_results=len(jobs), timeout=60.0)
@@ -27,7 +27,7 @@ def test_local_from_script_and_function():
     domain = ht.Domain({"x": {0, 1, 2, 3}, "y": [-1., 1.], "z": {"123", "abc"}}, seed=7)
     jobs = [Job(task="hypertunity/scheduling/tests/script.py::main",
                 args=(*domain.sample().as_namedtuple(),)) for _ in range(10)]
-    results = _run_jobs(jobs)
+    results = run_jobs(jobs)
     assert all([r.data == script.main(*j.args) for r, j in zip(results, jobs)])
 
 
@@ -37,7 +37,7 @@ def test_local_from_script_and_cmdline_args():
     jobs = [Job(task="hypertunity/scheduling/tests/script.py",
                 args=(*domain.sample().as_namedtuple(),),
                 meta={"binary": "python"}) for _ in range(10)]
-    results = _run_jobs(jobs)
+    results = run_jobs(jobs)
     assert all([r.data == script.main(*j.args) for r, j in zip(results, jobs)])
 
 
@@ -45,7 +45,7 @@ def test_local_from_script_and_cmdline_args():
 def test_local_from_fn():
     domain = ht.Domain({"x": [0., 1.]}, seed=7)
     jobs = [Job(task=square, args=(domain.sample(),)) for _ in range(10)]
-    results = _run_jobs(jobs)
+    results = run_jobs(jobs)
     assert all([r.data.value == square(*j.args).value for r, j in zip(results, jobs)])
 
 
@@ -64,7 +64,7 @@ def test_slurm_from_script():
                              args=(*sample.as_namedtuple(),),
                              output_file=f"{os.path.join(dirs[-1].name, 'results.pkl')}",
                              meta={"binary": "python", "resources": {"cpu": 1}}))
-    results = _run_jobs(jobs)
+    results = run_jobs(jobs)
     assert all([r.data == script.main(*j.args) for r, j in zip(results, jobs)])
     # clean-up the temporary dirs
     for d in dirs: